+++ RecipaediaEXLoader.cs
@@ -1,10 +1,11 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 using Game;
+using RecipaediaEX.Search;
 using RecipaediaEX.UI;
 
 namespace RecipaediaEX
@@ -25,6 +26,7 @@ namespace RecipaediaEX
             RecipesLoadManager.Initialize();
             RecipaediaEXManager.Initialize();
             RecipesCrafterManager.Initialize();
+            RecipaediaSearchIndex.Initialize();
             ScreensManager.m_screens["Recipaedia"] = new RecipaediaEXScreen();
             ScreensManager.m_screens["RecipaediaDescription"] = new RecipaediaEXDescriptionScreen();
             ScreensManager.m_screens["RecipaediaRecipes"] = new RecipaediaEXRecipesScreen();
@@ -40,5 +42,6 @@ namespace RecipaediaEX
         public static XElement RequestScreenFile(string screenName) => ContentManager.Get<XElement>($"RecipaediaEX/Screens/{screenName}");
         public static XElement RequestStyleFile(string screenName) => ContentManager.Get<XElement>($"RecipaediaEX/Styles/{screenName}");
         public static XElement RequestWidgetFile(string screenName) => ContentManager.Get<XElement>($"RecipaediaEX/Widgets/{screenName}");
+        public static XElement RequestDialogFile(string dialogName) => ContentManager.Get<XElement>($"RecipaediaEX/Dialogs/{dialogName}");
     }
 }