+++ ScreenExtra/Providers/BlocksCategoryProvider.cs
@@ -7,6 +7,8 @@ namespace RecipaediaEX.Implementation {
     public class BlocksCategoryProvider : IRecipaediaCategoryProvider {
         static List<IRecipaediaCategory> m_cachedCategories;
 
+        public static void InvalidateCache() => m_cachedCategories = null;
+
         public IEnumerable<IRecipaediaCategory> GetCategories() {
             if (m_cachedCategories == null) {
                 m_cachedCategories = BlocksManager.Categories